requirements.txt
Django==3.1.2
djangorestframework==3.12.1
gunicorn==20.0.4
PyMySQL==0.10.1
sqlparse==0.4.1
virtualenv==20.0.35
mysqlclient==2.0.1